test_files/requirements.txt
tabulate
dataclasses
XlsxWriter
pytest
pytest-html
compose_format
svg.charts